+++ CHANGELOG.md
@@ -227,6 +227,7 @@
 ### Feature
 
 - Lazy load image in blocks Image and HeroImage @mamico
+- Add Image with srcset and lazy loading using Plone scales @nzambello
 
 ### Bugfix